pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "argus-ops"
version = "0.4.0"
description = "AI-powered infrastructure monitoring CLI that detects, diagnoses, and remediates issues across Kubernetes, VMs, and bare metal"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Mason Kim", email = "ehehwnwjs5052@gmail.com"},
]
keywords = ["aiops", "monitoring", "kubernetes", "infrastructure", "devops", "sre", "observability"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Monitoring",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "click>=8.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "litellm>=1.0.0",
    "kubernetes>=28.0.0",
    "jinja2>=3.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "httpx>=0.27.0",
    "pytest-asyncio>=0.23.0",
    "bandit[toml]>=1.7.0",
    "safety>=3.0.0",
]
auth = [
    "bcrypt>=4.0.0",
]
web = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.27.0",
]

[project.scripts]
argus-ops = "argus_ops.cli:main"

[project.urls]
Homepage = "https://github.com/mason5052/argus-ops"
Repository = "https://github.com/mason5052/argus-ops"
Issues = "https://github.com/mason5052/argus-ops/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
argus_ops = ["web/templates/*.html", "ai/prompts/*.j2"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.bandit]
targets = ["src"]
skips = ["B101"]   # assert statements -- acceptable in non-security contexts
